// src/app/interfaces/IElectoresTmp.ts
export interface IElectoresTmp {
  amigo: string;
  apellidoCasado: string;
  apellidoMaterno: string;
  apellidoPaterno: string;
  barriada: string;
  barrio: string;
  calle: string;
  campaniaID: number;
  cedula: string;
  cedulaPareja: string;
  celular1: string;
  celular2: string;
  celularAmigo: string;
  centroVotacionID: number;
  comunidad: string;
  conyugePareja: string;
  corregimientoID: number;
  correo: string;
  createdAt: Date;
  distritoID: number;
  edad: number;
  edificio: string;
  estadoCivil: string;
  facebook: string;
  fechaNacimiento: Date;
  hijos: string;
  id: number;
  ingresoAproximado: string;
  instagram: string;
  lugarTrabajo: string;
  mesaNum: string;
  nivelEducativo: string;
  numCasaApto: string;
  otra: string;
  otrasReferencias: string;
  parentId: number;
  partidoID: number;
  pertenecePartido: string;
  primerNombre: string;
  profesionID: number;
  provinciaID: number;
  religionID: number;
  religionOtra: string;
  sector: string;
  segundoNombre: string;
  sexo: string;
  situacionLaboral: string;
  situacionVivienda: string;
  snapchat: string;
  telefono: string;
  telegram: string;
  tiktok: string;
  twitter: string;
  ubicacionLat: string;
  ubicacionLng: string;
  updatedAt: Date;
  userIdCreatedAt: number;
  userIdUpdatedAt: number;
  whatsapp: string;
  youtube: string;
  completeName: string;
}